Removing the management of AWS config file
The problem is that the config file is now a symlink pointing to the nix store, and this does not play well with docker.
In several places, we bind-mount the ~/.aws folder inside containers, but since now the config file in it is a symlink, docker cannot find the original file.
We'd have to bind-mount also the nix store, which does not make sense.