Fix vscode extensions setup
Having a mutable extension dir means we can update the extensions from the UI but we can't add extensions via the code, so I prefer the contrary.
It's not perfect and we need to find a way to easily find the sha of newer versions of extensions.

This also updates the ruby-lsp extension and remove the aws one which we don't use.