Update flake
There was a big change in nix-darwin around users: nix-darwin does not run as a user anymore, but as root, which means all user-focused configuration has to change.
The main thing for me is that the user activation scripts don't exist as-is anymore.
As a transition phase, a new primaryUser attribute has been set and all activationScripts are run by sudoing on the primary user.
IIUC, those are supposed to be moved to be managed by home-manager, but I couldn't find info on that yet.